pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-assist-context-engine"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu==1.7.4",
    "rank-bm25==0.2.2",
    "pyyaml==6.0.1",
    "tiktoken==0.5.1",
    "numpy==1.26.4",
    "scikit-learn==1.3.2",
    "pydantic==1.10.13",
]

[tool.pytest.ini_options]
addopts = "-q"
